Moved deprecated classes into an archive so they're out of the way.
