+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/archive/RoadrunnerSimpleSpecimen.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.archive;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
@@ -9,31 +9,32 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
+import org.firstinspires.ftc.teamcode.VerticalClaw;
 import org.firstinspires.ftc.teamcode.roadrunner.MecanumDrive;
 
-@Autonomous(name = "My Odometry Tester")
-@Deprecated
+// MUST TEST - New Bot
+@Autonomous(name = "Roadrunner Parking Specimen Autonomous")
 @Disabled
-public class MyOdometryTester extends LinearOpMode {
+@Deprecated
+public class RoadrunnerSimpleSpecimen extends LinearOpMode {
     @Override
     public void runOpMode() {
         MecanumDrive drive = new MecanumDrive(hardwareMap, new Pose2d(11.25, 35.04, Math.toRadians(270)));
+        Extension extension = new Extension(hardwareMap);
+        VerticalClaw claw = new VerticalClaw(hardwareMap);
+
+        Actions.runBlocking(new SequentialAction(
+                claw.hingeTo(VerticalClaw.HingePosition.UP),
+                claw.close()
+        ));
 
         telemetry.addData("Status", "Initialized");
         telemetry.update();
         waitForStart();
 
-        if (opModeIsActive()) {
-            TrajectoryActionBuilder traj = drive.actionBuilder(new Pose2d(10, 64, Math.toRadians(270)))
-                    .strafeTo(new Vector2d(10, 70))
-                    .strafeTo(new Vector2d(10, 58))
-                    .turn(Math.toRadians(-30));
+        TrajectoryActionBuilder toPark = drive.actionBuilder(new Pose2d(-16, 35.04, Math.toRadians(270)))
+                .strafeTo(new Vector2d(-40, 35));
 
-            Actions.runBlocking(
-                    new SequentialAction(
-                            traj.build()
-                    )
-            );
-        }
+        Actions.runBlocking(toPark.build());
     }
 }
